Created lago.prefix module
